@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "A general purpose Discord Bot",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "start": "cd build && node index.js",
     "dev": "cd build && NODE_ENV=dev nodemon -e js index.js",
@@ -25,36 +26,30 @@
     "delay": "1000"
   },
   "dependencies": {
-    "@discordjs/builders": "^0.6.0",
-    "@discordjs/opus": "^0.5.3",
-    "@discordjs/rest": "^0.1.0-canary.0",
-    "@discordjs/voice": "^0.5.5",
-    "@google-cloud/firestore": "^4.15.0",
-    "@types/node": "^14.17.19",
-    "bufferutil": "^4.0.4",
-    "chartjs-node-canvas": "^3.2.0",
-    "discord-api-types": "^0.22.0",
-    "discord.js": "^13.0.0",
+    "@discordjs/opus": "^0.9.0",
+    "@discordjs/voice": "^0.16.0",
+    "@google-cloud/firestore": "^4.15.1",
+    "@types/node": "^14.18.42",
+    "bufferutil": "^4.0.7",
+    "discord.js": "^14.9.0",
     "ffmpeg": "0.0.4",
-    "gaxios": "^4.3.0",
-    "libsodium-wrappers": "^0.7.9",
-    "node-cron": "^3.0.0",
-    "sodium": "^3.0.2",
-    "spotify-url-info": "^2.2.3",
-    "typescript": "^4.4.2",
-    "winston": "^3.3.3",
-    "winston-daily-rotate-file": "^4.5.5",
-    "ytdl-core": "^4.9.1",
+    "gaxios": "^4.3.3",
+    "libsodium-wrappers": "^0.7.11",
+    "node-cron": "^3.0.2",
+    "spotify-url-info": "^2.2.9",
+    "typescript": "^4.9.5",
+    "winston": "^3.8.2",
+    "winston-daily-rotate-file": "^4.7.1",
+    "ytdl-core": "^4.11.3",
     "ytdl-core-discord": "^1.3.1",
-    "ytsr": "^3.5.3"
+    "ytsr": "^3.8.1"
   },
   "devDependencies": {
-    "@types/chart.js": "^2.9.34",
-    "chai": "^4.3.4",
+    "chai": "^4.3.7",
     "cross-env": "^7.0.3",
     "eslint": "^7.32.0",
     "mocha": "^8.4.0",
-    "nodemon": "^2.0.13",
+    "nodemon": "^2.0.22",
     "opusscript": "0.0.8",
     "rewire": "^5.0.0"
   }
