@@ -1,15 +1,31 @@
-import Discord = require('discord.js');
+import { Client, GatewayIntentBits, ClientOptions, PermissionsBitField, RESTPostAPIChatInputApplicationCommandsJSONBody, SlashCommandBuilder } from 'discord.js'
+
+import { Collection } from 'discord.js';
 import { REST } from '@discordjs/rest';
 import { getVoiceConnection } from '@discordjs/voice';
-import { Routes } from 'discord-api-types/v9';
-import type { SlashCommandDataJSON } from '../typings/index';
-import { logger } from './log';
-import { getCommands, queryConfig } from './util';
+import { Routes } from 'discord-api-types/v10';
+import { queryConfig } from './util.js';
+import * as winston from 'winston';
+import { logger } from './log.js';
+
+class ClientExtend extends Client<true, any> {
+    public commands: Collection<string, SlashCommand>;
+    //public audioPlayers: Collection<string, AudioPlayer>;
+    public logger: winston.Logger;
+
+    constructor(options: ClientOptions) {
+        super(options);
+        this.logger = logger;
+        this.commands = new Collection();
+    }
+}
 
 const rest = new REST({ version: '9' });
-const client = new Discord.Client({
-    intents: ['GUILDS', 'GUILD_MESSAGES', 'GUILD_VOICE_STATES', 'GUILD_MESSAGE_REACTIONS', 'GUILD_MEMBERS']
+const client = new ClientExtend({
+    intents: [GatewayIntentBits.Guilds, GatewayIntentBits.GuildMessages, GatewayIntentBits.GuildVoiceStates, GatewayIntentBits.GuildMessageReactions, GatewayIntentBits.GuildMembers]
+    // BREAKING // intents: ['GUILDS', 'GUILD_MESSAGES', 'GUILD_VOICE_STATES', 'GUILD_MESSAGE_REACTIONS', 'GUILD_MEMBERS']
 });
+
 
 enum ClientEnums {
     AFK_TIMEOUT_MINUTES = 5,
@@ -19,10 +35,6 @@ enum ClientEnums {
     PROD_CLIENT_ID = '783886978974220338',
 }
 
-client.commands = new Map();
-client.musicQueueManager = new Discord.Collection;
-client.audioPlayers = new Map();
-client.logger = logger;
 
 /** Retrieve token from Firestore */
 queryConfig().then(config => {
@@ -37,6 +49,10 @@ queryConfig().then(config => {
     client.login(token);
 });
 
+
+import ExampleCommand from './commands/test.js'
+import { SlashCommand } from './commands/Command.js';
+
 /** Load commands after client initialized */
 client.once('ready', async () => {
     if (client.user)
@@ -46,22 +62,29 @@ client.once('ready', async () => {
             message: `Succesfully logged into ${client.user.tag}`
         });
 
-    const commands = await getCommands();
+    /* const commands = await getCommands();
     client.commands = commands;
 
-    let data: SlashCommandDataJSON[] = [];
+    let data: RESTPostAPIChatInputApplicationCommandsJSONBody[] = [];
     for (const [k, v] of commands) {
         data.push(v.data.toJSON())
     }
 
+    
+    */
+    let data: RESTPostAPIChatInputApplicationCommandsJSONBody[] = [];
+    
+    client.commands.set(ExampleCommand.name, ExampleCommand);
+    data.push(ExampleCommand.toJSON());
+
     if (process.env.NODE_ENV == 'dev') {
-        /** debugging guilds */
-        await rest.put(Routes.applicationGuildCommands(ClientEnums.TEST_CLIENT_ID, ClientEnums.TANK_SERVER_ID), { body: data });
+        // debugging guilds
+        //await rest.put(Routes.applicationGuildCommands(ClientEnums.TEST_CLIENT_ID, ClientEnums.TANK_SERVER_ID), { body: data });
         await rest.put(Routes.applicationGuildCommands(ClientEnums.TEST_CLIENT_ID, ClientEnums.DEV_SERVER_ID), { body: data });
     } else {
         await rest.put(Routes.applicationCommands(ClientEnums.PROD_CLIENT_ID), { body: data });
     }
-
+    
     await client.application?.commands.fetch();
     client.logger.log({
         level: 'info',
@@ -73,29 +96,38 @@ client.once('ready', async () => {
 /** Main command handler */
 client.on('interactionCreate', async interaction => {
     if (!interaction.isCommand() || !interaction.guild) return;
+    const interactionCommand = interaction.command;
     const command = interaction.commandName;
     try {
-        client.commands.get(command)!.execute(interaction);
-        client.logger.log({
-            level: 'verbose',
-            label: 'main',
-            guild: interaction.guild,
-            channel: interaction.channel,
-            user: interaction.user,
-            message: `Ran command ${command}`
-        });
+        
+        /* Check if command actually exists and has a function */
+        if (!(client.commands.get(command)) ||
+            (client.commands.get(command)?.execute == undefined)) {
+                console.log(client.commands.get(command));
+                throw `Command ${command} does not exist in client`;
+        } else {
+            client.commands.get(command)!.execute(interaction);
+            client.logger.log({
+                level: 'verbose',
+                label: 'main',
+                guild: interaction.guild,
+                channel: interaction.channel,
+                user: interaction.user,
+                message: `Ran command ${command}`
+            });
+        }
     } catch (e) {
         client.logger.log({
             level: 'error',
             label: 'main',
-            message: e.stack
+            message: e.stack ?? e
         });
     }
 });
 
 /** Remove all commands and kill process */
 client.on('messageCreate', async message => {
-    const hasPermissions = message.member?.permissions.has(Discord.Permissions.FLAGS.ADMINISTRATOR);
+    const hasPermissions = message.member?.permissions.has(PermissionsBitField.Flags.Administrator);
     if (hasPermissions && message.content.startsWith('.refresh')) {
         client.logger.log({
             level: 'info',
@@ -106,25 +138,9 @@ client.on('messageCreate', async message => {
             for (const cmd of cmds) await client.application?.commands.delete(cmd[1]);
         }).finally(() => process.exit(-1));
     }
-
-    // quote channel cleaning up
-    if (message.channel.id == '747290626038431875' && !message.author.bot) {
-        if (!message.content.match(/^['">]/gi)) {
-            let response = await message.reply({ content: `\`${message.content}\` doesn't appear to be quoting anything. Perhaps use a reaction instead?\nThis message will delete itself in 10 seconds..`});
-            setTimeout(() => {
-                response.delete();
-                if (message && message.deletable) message.delete();
-            }, 10000);
-        }
-    }
-
-    if (message.channel.id == '951924418408505434') {
-        // delete message after 1 minute
-        setTimeout(() => { message.delete(); }, 60 * 1000);
-    }
 });
 
-/** AFK Timeout (5 minutes) */
+/** VC AFK Timeout (5 minutes) */
 client.on('voiceStateUpdate', (pre) => {
     let connection = getVoiceConnection(pre.guild.id),
         queue = client.musicQueueManager.get(pre.guild.id)!;
